+++ include/talmech/evaluator.h
@@ -1,0 +1,12 @@
+#ifndef _TALMECH_EVALUATOR_H_
+#define _TALMECH_EVALUATOR_H_
+
+namespace talmech
+{
+class Evaluator
+{
+public:
+  Evaluator();
+};
+
+#endif // _TALMECH_EVALUATOR_H_