+++ src/talmech/robot.cpp
@@ -1,0 +1,6 @@
+#include "talmech/robot.h"
+
+namespace talmech
+{
+
+}