+++ CHANGELOG.rst
@@ -7,3 +7,15 @@ Changelog for package talmech
 * created README.md file
 * created package
 * Contributors: adrianohrl
+
+Forthcoming
+-----------
+* created bidder controller related classes
+* created auctioneer machine state related classes
+* added UML diagrams
+* Contributors: adrianohrl
+
+0.0.2 (2018-01-25)
+------------------
+* configured package
+* Contributors: adrianohrl